app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.0
cryptography==42.0.7
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.9
ddtrace==2.8.5
pydantic==2.7.1
pysnmp==4.4.12
psutil==5.9.8
requests==2.31.0
